MAUI-1267: Code cleanups
