+++ packages/central-server/src/apiV2/entities/index.js
@@ -3,5 +3,7 @@
  * Copyright (c) 2017 - 2021 Beyond Essential Systems Pty Ltd
  */
 
-export { DeleteEntityHandler } from './DeleteEntityHandler';
+export { DeleteEntity } from './DeleteEntity';
 export { EditEntity } from './EditEntity';
+export { GETEntities } from './GETEntities';
+export { assertEntityPermissions } from './assertEntityPermissions';
